demo python version upgrade
